+++ src/www/components/popup/PopupComp.scss
@@ -1,0 +1,171 @@
+@import "../../vars.scss";
+
+.PopupTrigger {
+  background-color: rgba($color: rgb(221, 191, 191), $alpha: 0.1);
+  border: 1px solid rgba($color: black, $alpha: 0.25);
+  border-radius: 0.4rem;
+  cursor: pointer;
+  outline: none;
+  padding: 0.5rem;
+  width: 12%;
+}
+
+.popup-content {
+  border-radius: 1rem;
+  max-width: 40rem;
+
+  .close {
+    background: #ffffff;
+    border: 1px solid #cfcece;
+    border-radius: 18px;
+    cursor: pointer;
+    font-size: 24px;
+    line-height: 15px;
+    outline: none;
+    padding: 2.5px 5px;
+    position: absolute;
+    right: 0px;
+    top: 0px;
+  }
+
+  .scrollarea {
+    height: 37rem;
+
+    .scrollarea-content {
+      > p {
+        color: red;
+        margin: 0rem;
+      }
+
+      > form {
+        display: flex;
+        flex-direction: column;
+        margin: 1rem 3rem;
+
+        > input {
+          height: 1.5rem;
+        }
+
+        > textarea {
+          height: 5rem;
+          resize: none;
+        }
+
+        > input,
+        textarea {
+          border: 1px solid $softBlack;
+          border-radius: 0.6rem;
+          box-shadow: 1px 1px $softBlack;
+          margin-bottom: 1rem;
+          outline: none;
+          padding: 0.3rem;
+        }
+
+        > div {
+          margin-bottom: 1rem;
+
+          > p {
+            font-weight: bold;
+            text-align: left;
+            margin: 0.4rem 0;
+          }
+
+          .scrollarea {
+            height: 8rem;
+
+            .scrollarea-content {
+              display: flex;
+              flex-wrap: wrap;
+              text-align: left;
+              flex-direction: column;
+
+              > label {
+                display: block;
+                position: relative;
+                padding-left: 40px;
+                margin-bottom: 0.3rem;
+                cursor: pointer;
+                font-size: 1rem;
+                user-select: none;
+                text-align: start;
+
+                &:hover {
+                  color: lightskyblue;
+                }
+
+                input {
+                  position: absolute;
+                  opacity: 0;
+                  cursor: pointer;
+                  height: 0;
+                  width: 0;
+                }
+
+                .checkmark {
+                  position: absolute;
+                  top: 0;
+                  left: 5px;
+                  height: 20px;
+                  width: 20px;
+                  background-color: rgba($color: white, $alpha: 0.75);
+                  border: 1px solid $softBlack;
+
+                  &:after {
+                    content: "";
+                    position: absolute;
+                    display: none;
+                  }
+                }
+              }
+
+              &:hover input ~ .checkmark {
+                background-color: rgba($color: white, $alpha: 0.25);
+                transition: 0.2s ease-in-out;
+              }
+
+              input:checked ~ .checkmark {
+                background-color: lightskyblue;
+              }
+
+              input:checked ~ .checkmark:after {
+                display: block;
+              }
+
+              .checkmark:after {
+                left: 7px;
+                top: 3px;
+                width: 3px;
+                height: 8px;
+                border: solid white;
+                border-width: 0 3px 3px 0;
+                transform: rotate(45deg);
+              }
+            }
+          }
+
+          .fileContainer {
+            margin: 0rem;
+            padding: 0rem;
+
+            .deleteImage {
+              background: lightskyblue;
+              font-size: 20px;
+              line-height: 23px;
+              height: 20px;
+              width: 20px;
+            }
+          }
+        }
+
+        .SubmitButton {
+          align-self: center;
+          cursor: pointer;
+          height: 2rem;
+          margin: 0rem;
+          outline: none;
+          width: 10rem;
+        }
+      }
+    }
+  }
+}
