+++ src/www/containers/teamspage/TeamsPage.jsx
@@ -1,19 +1,14 @@
 import "./TeamsPage.scss";
 
 import React, { Component } from "react";
-import { Redirect } from "react-router-dom";
-import Popup from "reactjs-popup";
-import Input from "../../components/input/Input";
-import axios from "axios";
 import { withTranslation } from "react-i18next";
 
 import FilterList from "../../components/filterlist/FilterList";
 import Pagination from "../../components/pagination/Pagination";
 import SearchBar from "../../components/searchbar/SearchBar";
+import PopupComp from "../../components/popup/PopupComp";
 import queryString from "query-string";
 import { withRouter } from "react-router-dom";
-import ScrollArea from "react-scrollbar";
-import ImageUploader from "react-images-upload";
 
 const categories = require("./categories.json");
 const teams = require("./teams.json");
@@ -24,66 +19,12 @@ export default withRouter(
       state = {
         selectedCategories: [],
         query: "",
-        teamName: "",
-        description: "",
-        teamCategories: [],
-        teamPicture: "",
-        serverResponse: "",
-        redirect: false,
-      };
-
-      createTeam = (name, description, categories, photo) => {
-        axios
-          .post("/api/teams/create", {
-            name,
-            description,
-            categories,
-            photo,
-          })
-          .then((res) => this.handleResponse(res.data))
-          .catch((err) => this.handleResponse(err.response.data));
       };
 
       componentDidMount = () => {
         const { query } = queryString.parse(this.props.location.search);
         if (query) {
           this.setState({ query });
-        }
-      };
-
-      handleResponse = (data) => {
-        const { t } = this.props;
-        if (data.error) {
-          if (data.error.includes("Team name")) {
-            this.setState({
-              serverResponse: t("createteam.nameerror"),
-            });
-          } else if (data.error.includes("Authentication")) {
-            this.setState({
-              serverResponse: t("createteam.authentication"),
-            });
-          }
-        } else {
-          this.setState({
-            redirect: true,
-          });
-        }
-      };
-
-      handleSubmit = (event) => {
-        event.preventDefault();
-        const {
-          teamName,
-          description,
-          teamCategories,
-          teamPicture,
-        } = this.state;
-
-        if (teamName !== "" && description !== "") {
-          this.setState({
-            teamNameHighlight: false,
-          });
-          this.createTeam(teamName, description, teamCategories, teamPicture);
         }
       };
 
@@ -123,13 +64,6 @@ export default withRouter(
       };
 
       render() {
-        const { t } = this.props;
-
-        if (this.state.redirect) {
-          // TODO change `id` to the _id sent from server
-          return <Redirect to="/teams/id" />;
-        }
-
         return (
           <div className="TeamsPage">
             <div>
@@ -138,105 +72,7 @@ export default withRouter(
                 value={this.state.query}
                 onChange={(e) => this.setState({ query: e.target.value })}
               />
-              <Popup
-                trigger={<button>{t("createteam.createteam")}</button>}
-                closeOnDocumentClick={false}
-                modal
-              >
-                {(close) => (
-                  <ScrollArea>
-                    <button
-                      className="close"
-                      onClick={() => {
-                        close();
-                        this.setState({
-                          serverResponse: "",
-                        });
-                      }}
-                    >
-                      &times;
-                    </button>
-                    <h2>{t("createteam.createteam")}</h2>
-                    <p children={this.state.serverResponse} />
-                    <form
-                      method="post"
-                      action="api/teams/create"
-                      onSubmit={this.handleSubmit}
-                    >
-                      <Input
-                        name="teamName"
-                        onChange={(e) =>
-                          this.setState({ teamName: e.target.value.trim() })
-                        }
-                        type="text"
-                        placeholder={t("createteam.teamname")}
-                      ></Input>
-                      <textarea
-                        name="description"
-                        onChange={(e) =>
-                          this.setState({ description: e.target.value.trim() })
-                        }
-                        required
-                        type="text"
-                        placeholder={t("createteam.description")}
-                      ></textarea>
-                      <div>
-                        <p>{t("filterlist.categories")}</p>
-                        <ScrollArea>
-                          {Object.keys(categories).map((c) => {
-                            const category = categories[c];
-                            return (
-                              <label key={c}>
-                                {t(`categories.${category.name.toLowerCase()}`)}
-                                <input
-                                  type="checkbox"
-                                  onChange={(event) => {
-                                    if (event.target.checked) {
-                                      const categories = this.state.teamCategories.concat(
-                                        category.name
-                                      );
-                                      this.setState({
-                                        teamCategories: categories,
-                                      });
-                                    } else {
-                                      const filtered = this.state.teamCategories.filter(
-                                        (c) => c !== category.name
-                                      );
-
-                                      this.setState({
-                                        teamCategories: filtered,
-                                      });
-                                    }
-                                  }}
-                                />
-                                <span className="checkmark"></span>
-                              </label>
-                            );
-                          })}
-                        </ScrollArea>
-                      </div>
-                      <ImageUploader
-                        withIcon={true}
-                        buttonText="Choose image"
-                        onChange={(pic) => {
-                          this.setState({
-                            teamPicture: pic,
-                          });
-                        }}
-                        imgExtension={[".jpg", ".gif", ".png", ".gif"]}
-                        maxFileSize={5242880}
-                        singleImage={true}
-                        withPreview={true}
-                      />
-                      <input
-                        className="SubmitButton"
-                        type="submit"
-                        value={t("submit")}
-                      />
-                    </form>
-                  </ScrollArea>
-                )}
-              </Popup>
+              <PopupComp categories={categories} />
             </div>
             <span>
               <FilterList
